Move sleep to lodestar-utils
