+++ packages/lodestar-utils/package.json
@@ -37,6 +37,7 @@
   "types": "lib/index.d.ts",
   "dependencies": {
     "@chainsafe/ssz": "^0.6.13",
+    "abort-controller": "^3.0.0",
     "bigint-buffer": "^1.1.5",
     "chalk": "^2.4.2",
     "js-yaml": "^3.13.1",
