+++ packages/lodestar-utils/src/errors.ts
