+++ packages/lodestar-utils/src/index.ts
@@ -1,8 +1,10 @@
 export * from "./assert";
+export * from "./errors";
 export * from "./math";
 export * from "./bytes";
 export * from "./yaml";
 export * from "./objects";
+export * from "./sleep";
 export * from "./sort";
 export * from "./logger";
 export * from "./verifyMerkleBranch";
