+++ packages/lodestar/test/e2e/network/network.test.ts
@@ -6,11 +6,11 @@ import {generateEmptySignedBlock} from "../../utils/block";
 import {ILogger, WinstonLogger} from "@chainsafe/lodestar-utils/lib/logger";
 import {INetworkOptions} from "../../../src/network/options";
 import {BeaconMetrics} from "../../../src/metrics";
-import {sleep} from "../../../src/util/sleep";
 import Libp2p from "libp2p";
 import sinon, {SinonStubbedInstance} from "sinon";
 import {GossipMessageValidator} from "../../../src/network/gossip/validator";
 import {Attestation, SignedBeaconBlock} from "@chainsafe/lodestar-types";
+import {sleep} from "@chainsafe/lodestar-utils";
 import {generateState} from "../../utils/state";
 import {MockBeaconChain} from "../../utils/mocks/chain/chain";
 import {IBeaconChain} from "../../../src/chain";
